# Part2_Infrastructure/tools/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckResult": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "limit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "observed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "passed",
          "detail"
        ],
        "title": "CheckResult",
        "type": "object"
      },
...
      "Fill": {
        "properties": {
          "fee_usd": {
            "title": "Fee Usd",
            "type": "number"
          },
          "notional": {
            "title": "Notional",
            "type": "number"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "number"
          },
          "simulated": {
            "default": true,
            "title": "Simulated",
            "type": "boolean"
          },
          "slippage_bps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slippage Bps"
          },
          "venue": {
            "title": "Venue",
            "type": "string"
          }
        },
        "required": [
          "price",
          "quantity",
          "notional",
          "fee_usd",
          "slippage_bps",
          "venue"
        ],
        "title": "Fill",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderRequest": {
        "properties": {
          "client_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Order Id"
          },
          "limit_price": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "notional": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notional"
          },
          "order_type": {
            "default": "MARKET",
            "enum": [
              "MARKET",
              "LIMIT"
            ],
            "title": "Order Type",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "side": {
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Side",
            "type": "string"
          },
          "strategy": {
            "default": "manual",
            "title": "Strategy",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "side"
        ],
        "title": "OrderRequest",
        "type": "object"
      },
...
      "RiskDecision": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckResult"
            },
            "title": "Checks",
            "type": "array"
          },
          "client_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Order Id"
          },
          "fill": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Fill"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "default": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "notional": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notional"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rejected_by": {
            "items": {
              "type": "string"
            },
            "title": "Rejected By",
            "type": "array"
          },
          "side": {
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Side",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "client_order_id",
          "accepted",
          "symbol",
          "side",
          "quantity",
          "notional",
          "checks",
          "timestamp"
        ],
        "title": "RiskDecision",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Unified execution-quality, pre-trade-risk and research gateway. Module A: cross-venue L2 TCA \u00b7 Module B: pre-trade risk & kill switch \u00b7 Module C: asynchronous parametric backtesting.",
    "title": "AlphaEngine Trading Automation \u2014 Execution Gateway & Risk Portal",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "description": "The only way an order reaches a venue. Returns the full check vector for\nboth accepted and rejected orders \u2014 a rejection is a result, not an error.",
        "operationId": "submit_order_api_orders_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-AlphaEngine-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Alphaengine-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskDecision"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Order",
        "tags": [
          "B \u00b7 Risk"
        ]
      }
    },
...
  }
}